+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TBD
+### Changed
+* Improved error logging when extraction job fails or is no longer present.
+
 # 1.0.1 - 2018-02-28
 ### Changed
-* Project moved into Github
+* Project moved into Github.