+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # TBD
 ### Changed
 * Improved error logging when extraction job fails or is no longer present.
+### Fixed
+* Explicitly fail table replication when the BigQuery source is a view rather than failing silently. See [#47](https://github.com/HotelsDotCom/circus-train-bigquery/issues/5).
 
 # 1.0.1 - 2018-02-28
 ### Changed