+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# TBD
+# 2.0.0 - 2018-07-13
 ### Changed
+* Updated project to be compatible with Circus Train 12.0.0 and up
 * Improved error logging when extraction job fails or is no longer present.
 * Extract table during replication rather than in listener to prevent silent failure. See [#5](https://github.com/HotelsDotCom/circus-train-bigquery/issues/5).